Fixes #1
Adds the language code and locale for the current site language to the `lang` attribute of the `html` tag. `str_replace` here is used to replace the underscore to a dash to comply with the W3C Recommendation defined by a 2 primary code (ISO639) for language and 2 letter subcode (ISO3166) for country.

Example: `fr-CA` would represent French language spoken in Canada.